+++ lab/ex-summary.R
@@ -1,0 +1,3 @@
+library(taxlist)
+
+taxlist:::overview_taxlist(object, units, check_validity)