+++ lab/test-df2taxlist.R
@@ -1,0 +1,13 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+library(taxlist)
+
+Cyperus <- read.csv(system.file("cyperus", "names.csv", package = "taxlist"),
+    stringsAsFactors = FALSE
+)
+Cyperus$AcceptedName <- !Cyperus$SYNONYM
+
+new_tax <- df2taxlist(Cyperus)